restocking/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==3.1.3
Flask-Migrate==4.0.5
gunicorn==21.2.0
psycopg[binary]==3.2.9
alembic==1.13.1
flask-login==0.6.2